refactor(map): wave 5 slices B–D — gen/view splits and 350 NCL ratchet (#4297)
Extract view-builder assembly, generator pass adapters, and region-grow
phase helpers into concern-sized libraries; lower headroom ceiling 400→350
with empty grandfather list. All 346 map tests pass.

Refs #4297
